requirements.txt
asn1crypto==1.5.1
attrs==21.4.0
awscli==1.25.30
boto3==1.24.28
botocore==1.27.30
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.4
colorful==0.5.4
cryptography==36.0.2
docutils==0.16
idna==3.3
iniconfig==1.1.1
jmespath==1.0.1
oscrypto==1.3.0
packaging==21.3
pluggy==1.0.0
prettyprinter==0.18.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
snowflake-connector-python==2.7.9
sqlvalidator==0.0.19
tomli==2.0.1
urllib3==1.26.10